Move RageSurface to RageUtil/Graphics/Surface
