+++ src/Etterna/Models/Misc/ImageCache.cpp
@@ -4,12 +4,12 @@
 #include "RageUtil/File/RageFileManager.h"
 #include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "Core/Services/Locator.hpp"
-#include "RageUtil/Graphics/RageSurface.h"
-#include "RageUtil/Graphics/RageSurfaceUtils.h"
-#include "RageUtil/Graphics/RageSurfaceUtils_Dither.h"
-#include "RageUtil/Graphics/RageSurfaceUtils_Palettize.h"
-#include "RageUtil/Graphics/RageSurfaceUtils_Zoom.h"
-#include "RageUtil/Graphics/RageSurface_Load.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Dither.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Palettize.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Zoom.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Load.h"
 #include "RageUtil/Graphics/RageTexture.h"
 #include "RageUtil/Graphics/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
