+++ src/Etterna/Models/Songs/Song.cpp
@@ -2,7 +2,7 @@
 #include "Etterna/Singletons/GameManager.h"
 #include "Core/Services/Locator.hpp"
 #include "RageUtil/Sound/RageSoundReader_FileReader.h"
-#include "RageUtil/Graphics/RageSurface_Load.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Load.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Song.h"
 #include "SongCacheIndex.h"
@@ -12,7 +12,7 @@
 #include "Etterna/Models/Misc/ImageCache.h"
 #include "Etterna/Actor/Base/Sprite.h"
 #include "RageUtil/File/RageFileManager.h"
-#include "RageUtil/Graphics/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
 #include "RageUtil/Graphics/RageTextureManager.h"
 #include "SongUtil.h"
 #include "Etterna/Singletons/SongManager.h"
