+++ src/RageUtil/CMakeLists.txt
@@ -54,16 +54,16 @@ list(APPEND SMDATA_RAGE_GRAPHICS_SRC
   "Graphics/Display/OpenGL/ShaderGL.cpp"
   "Graphics/Display/OpenGL/GLHelper.cpp"
   "Graphics/RageModelGeometry.cpp"
-  "Graphics/RageSurface.cpp"
-  "Graphics/RageSurface_Load.cpp"
-  "Graphics/RageSurface_Load_XPM.cpp"
-  "Graphics/RageSurface_Save_BMP.cpp"
-  "Graphics/RageSurface_Save_JPEG.cpp"
-  "Graphics/RageSurface_Save_PNG.cpp"
-  "Graphics/RageSurfaceUtils.cpp"
-  "Graphics/RageSurfaceUtils_Dither.cpp"
-  "Graphics/RageSurfaceUtils_Palettize.cpp"
-  "Graphics/RageSurfaceUtils_Zoom.cpp"
+  "Graphics/Display/Surface/RageSurface.cpp"
+  "Graphics/Display/Surface/RageSurface_Load.cpp"
+  "Graphics/Display/Surface/RageSurface_Load_XPM.cpp"
+  "Graphics/Display/Surface/RageSurface_Save_BMP.cpp"
+  "Graphics/Display/Surface/RageSurface_Save_JPEG.cpp"
+  "Graphics/Display/Surface/RageSurface_Save_PNG.cpp"
+  "Graphics/Display/Surface/RageSurfaceUtils.cpp"
+  "Graphics/Display/Surface/RageSurfaceUtils_Dither.cpp"
+  "Graphics/Display/Surface/RageSurfaceUtils_Palettize.cpp"
+  "Graphics/Display/Surface/RageSurfaceUtils_Zoom.cpp"
   "Graphics/RageTexture.cpp"
   "Graphics/RageTextureID.cpp"
   "Graphics/RageTextureManager.cpp"
@@ -79,16 +79,16 @@ list(APPEND SMDATA_RAGE_GRAPHICS_HPP
   "Graphics/Display/OpenGL/ShaderGL.hpp"
   "Graphics/Display/OpenGL/GLHelper.hpp"
   "Graphics/RageModelGeometry.h"
-  "Graphics/RageSurface.h"
-  "Graphics/RageSurface_Load.h"
-  "Graphics/RageSurface_Load_XPM.h"
-  "Graphics/RageSurface_Save_BMP.h"
-  "Graphics/RageSurface_Save_JPEG.h"
-  "Graphics/RageSurface_Save_PNG.h"
-  "Graphics/RageSurfaceUtils.h"
-  "Graphics/RageSurfaceUtils_Dither.h"
-  "Graphics/RageSurfaceUtils_Palettize.h"
-  "Graphics/RageSurfaceUtils_Zoom.h"
+  "Graphics/Display/Surface/RageSurface.h"
+  "Graphics/Display/Surface/RageSurface_Load.h"
+  "Graphics/Display/Surface/RageSurface_Load_XPM.h"
+  "Graphics/Display/Surface/RageSurface_Save_BMP.h"
+  "Graphics/Display/Surface/RageSurface_Save_JPEG.h"
+  "Graphics/Display/Surface/RageSurface_Save_PNG.h"
+  "Graphics/Display/Surface/RageSurfaceUtils.h"
+  "Graphics/Display/Surface/RageSurfaceUtils_Dither.h"
+  "Graphics/Display/Surface/RageSurfaceUtils_Palettize.h"
+  "Graphics/Display/Surface/RageSurfaceUtils_Zoom.h"
   "Graphics/RageTexture.h"
   "Graphics/RageTextureID.h"
   "Graphics/RageTextureManager.h"
