+++ src/RageUtil/Graphics/Display/OpenGL/RageDisplay_OGL.cpp
@@ -7,8 +7,8 @@
 #include "RageUtil/File/RageFile.h"
 #include "Core/Services/Locator.hpp"
 #include "RageUtil/Misc/RageMath.h"
-#include "RageUtil/Graphics/RageSurface.h"
-#include "RageUtil/Graphics/RageSurfaceUtils.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils.h"
 #include "RageUtil/Graphics/RageTextureManager.h"
 #include "RageUtil/Misc/RageTypes.h"
 #include "RageUtil/Utils/RageUtil.h"
