+++ src/RageUtil/Graphics/Display/RageDisplay.cpp
@@ -7,11 +7,11 @@
 #include "RageUtil/File/RageFile.h"
 #include "Core/Services/Locator.hpp"
 #include "RageUtil/Misc/RageMath.h"
-#include "RageUtil/Graphics/RageSurface.h"
-#include "RageUtil/Graphics/RageSurfaceUtils_Zoom.h"
-#include "RageUtil/Graphics/RageSurface_Save_BMP.h"
-#include "RageUtil/Graphics/RageSurface_Save_JPEG.h"
-#include "RageUtil/Graphics/RageSurface_Save_PNG.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Zoom.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Save_BMP.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Save_JPEG.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Save_PNG.h"
 #include "RageUtil/Misc/RageTimer.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Etterna/Screen/Others/Screen.h"
