+++ src/RageUtil/Graphics/RageBitmapTexture.cpp
@@ -2,11 +2,11 @@
 #include "RageBitmapTexture.h"
 #include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "Core/Services/Locator.hpp"
-#include "RageSurface.h"
-#include "RageSurfaceUtils.h"
-#include "RageSurfaceUtils_Dither.h"
-#include "RageSurfaceUtils_Zoom.h"
-#include "RageSurface_Load.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Dither.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Zoom.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Load.h"
 #include "RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Etterna/Globals/StepMania.h"
