+++ src/arch/MovieTexture/MovieTexture_Generic.cpp
@@ -1,10 +1,10 @@
 #include "Etterna/Globals/global.h"
 #include "MovieTexture_Generic.h"
-#include "RageUtil/Graphics/RageSurfaceUtils.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils.h"
 #include "Etterna/Singletons/PrefsManager.h"
 #include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "Core/Services/Locator.hpp"
-#include "RageUtil/Graphics/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
 #include "RageUtil/Graphics/RageTextureManager.h"
 #include "RageUtil/Graphics/RageTextureRenderTarget.h"
 #include "RageUtil/Utils/RageUtil.h"
