+++ src/arch/MovieTexture/MovieTexture_Null.cpp
@@ -4,7 +4,7 @@
 #include "RageUtil/Utils/RageUtil.h"
 #include "Core/Services/Locator.hpp"
 #include "MovieTexture_Null.h"
-#include "RageUtil/Graphics/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
 
 class MovieTexture_Null : public RageMovieTexture
 {
