+++ src/archutils/Win32/WindowIcon.cpp
@@ -1,9 +1,9 @@
 #include "Etterna/Globals/global.h"
 #include "archutils/Win32/WindowIcon.h"
 #include "Core/Services/Locator.hpp"
-#include "RageUtil/Graphics/RageSurface.h"
-#include "RageUtil/Graphics/RageSurfaceUtils.h"
-#include "RageUtil/Graphics/RageSurface_Load.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils.h"
+#include "RageUtil/Graphics/Display/Surface/RageSurface_Load.h"
 #include "archutils/Win32/ErrorStrings.h"
 
 #include <wingdi.h>
